requirements.txt
opencv-python
opencv-python-headless
tensorflow
tensorflow-plot
cython
pillow
torch
torchvision
opencv-python
numpy
shapely
tqdm
pandas
PyYAML
seaborn
scipy
imutils
facenet_pytorch
python-telegram-bot